@@ -6,8 +6,9 @@
     <version>1.0-SNAPSHOT</version>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>15</maven.compiler.source>
+        <maven.compiler.target>15</maven.compiler.target>
+        <project.main.class>ru.gb.java2.chat.client.ClientChat</project.main.class>
     </properties>
     <dependencies>
         <dependency>
@@ -22,7 +23,7 @@
         </dependency>
         <dependency>
             <groupId>NetworkChat</groupId>
-            <artifactId>NetworkChatClientServer</artifactId>
+            <artifactId>ChatClientServer</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
@@ -33,7 +34,7 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.0</version>
                 <configuration>
-                    <release>11</release>
+                    <release>15</release>
                 </configuration>
             </plugin>
             <plugin>
@@ -43,6 +44,19 @@
                 <configuration>
                     <mainClass>ru.gb.java2.chat.client.ClientChat</mainClass>
                 </configuration>
+                <executions>
+                    <execution>
+                        <!-- Configuration for debugging -->
+                        <id>debug</id>
+                        <configuration>
+                            <options>
+                                <option>-agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=*:8000</option>
+                            </options>
+                            <mainClass>${project.main.class}</mainClass>
+                        </configuration>
+                    </execution>
+                </executions>
+
             </plugin>
         </plugins>
     </build>
