@@ -3,7 +3,7 @@ package ru.gb.java2.chat.server;
 import ru.gb.java2.chat.server.chat.MyServer;
 
 public class ServerApp {
-    private static final int DEFAULT_PORT = 8189;
+    private static final int DEFAULT_PORT = 8182;
 
     public static void main(String[] args) {
         int port = DEFAULT_PORT;
